requirements.txt
aiogram==2.25.1
python_dotenv==1.0.0
beautifulsoup4
requests